requirements.txt
aiosqlite==0.21.0
alembic==1.17.0
annotated-types==0.7.0
anyio==4.11.0
apprise==1.9.5
asgi-lifespan==2.1.0
async-timeout==5.0.1
asyncpg==0.29.0
attrs==25.4.0
Authlib==1.6.5
blinker==1.9.0
boto3
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.3
coolname==2.2.0
croniter==5.0.1
cryptography==46.0.3
cycler==0.12.1
cyclopts==4.0.0
databricks-sdk==0.69.0
dateparser==1.2.2
dnspython==2.8.0
docker==7.1.0
docstring_parser==0.17.0
docutils==0.22.2
durationpy==0.10
dvc[s3]==3.58.0
email-validator==2.3.0
evidently
exceptiongroup==1.3.0
fastapi==0.119.1
fastmcp==2.12.5
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.60.1
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.41.1
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
graphviz==0.21
greenlet==3.2.4
griffe==1.14.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
humanize==4.14.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.4.5
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
keras==3.8.0
kiwisolver==1.4.9
kubernetes==31.0.0
lazy-object-proxy==1.12.0
Mako==1.3.10
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mcp==1.16.0
mdurl==0.1.2
mlflow==3.5.0
mlflow-skinny==3.5.0
mlflow-tracing==3.5.0
more-itertools==10.8.0
numpy==2.0.2
oauthlib==3.3.1
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opentelemetry-api==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.3
packaging==25.0
pandas==2.3.3
parse==1.20.2
pathable==0.4.4
pathspec==0.12.1
pendulum==2.1.2
pillow==12.0.0
prefect==2.20.23
prometheus-client
protobuf==5.28.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
pyparsing==3.2.5
pyperclip==1.11.0
pytest
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
pytzdata==2020.1
PyYAML==6.0.3
readchar==4.2.1
referencing==0.36.2
regex==2025.10.23
requests==2.32.5
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==14.2.0
rich-rst==1.3.2
rpds-py==0.27.1
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
scikit-learn==1.7.2
scipy==1.16.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlparse==0.5.3
sse-starlette==3.0.2
starlette==0.48.0
tensorflow==2.18.0
text-unidecode==1.3
threadpoolctl==3.6.0
toml==0.10.2
typer==0.17.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.38.0
websocket-client==1.9.0
websockets==13.1
Werkzeug==3.1.1
zipp==3.23.0
